+++ src/main/java/calculator/service/arithmetic/Arithmetic.java
@@ -1,4 +1,4 @@
-package service.arithmetic;
+package calculator.service.arithmetic;
 
 public interface Arithmetic {
