+++ src/main/java/calculator/service/arithmetic/Divide.java
@@ -1,4 +1,4 @@
-package service.arithmetic;
+package calculator.service.arithmetic;
 
 public class Divide implements Arithmetic {
