+++ src/main/java/calculator/service/arithmetic/Multiply.java
@@ -1,4 +1,4 @@
-package service.arithmetic;
+package calculator.service.arithmetic;
 
 public class Multiply implements Arithmetic {
