+++ src/main/java/calculator/service/arithmetic/Operator.java
@@ -1,4 +1,4 @@
-package service.arithmetic;
+package calculator.service.arithmetic;
 
 import java.util.function.BiFunction;
