+++ src/main/java/calculator/service/arithmetic/Plus.java
@@ -1,4 +1,4 @@
-package service.arithmetic;
+package calculator.service.arithmetic;
 
 public class Plus implements Arithmetic {
