+++ src/main/java/calculator/service/arithmetic/Subtract.java
@@ -1,4 +1,4 @@
-package service.arithmetic;
+package calculator.service.arithmetic;
 
 public class Subtract implements Arithmetic {
