+++ src/main/java/calculator/utility/parser/InputParser.java
@@ -1,4 +1,4 @@
-package parser;
+package calculator.utility.parser;
 
 import java.util.ArrayList;
 import java.util.List;
