+++ src/test/java/calculator/CalculatorServiceTest.java
@@ -1,16 +1,14 @@
 package calculator;
 
 import org.junit.jupiter.api.Test;
-import service.calculator.Calculator;
-
-import java.util.List;
+import calculator.service.CalculatorService;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class CalculatorTest {
+public class CalculatorServiceTest {
 
     @Test
     public void 문자열_계산기() {
-        assertThat(Calculator.calculate("2 + 3 * 4 / 2")).isEqualTo(10);
+        assertThat(CalculatorService.calculate("2 + 3 * 4 / 2")).isEqualTo(10);
     }
 }
